Implement enable/disable echo for when the user is entering their password
This was a real pain. There aren't many libraries about this out there, and RFCs 854 and 857
specificy what should happen between the server/client communication, they don't talk about
which specific bytes to use or exactly which ACK the receiver should use.

I ended up cobbling together the correct byte sequences from multiple sources, but Ruby
didn't _really_ want to deal with those either. Finally, while I know the enable/disable
bytes are correct, the ACKs are all over the place, so we may have to add more of those
to the list in the future.